refactor(projects/products/unstable/code_forge_engine,projects/products/unstable/code_forge_engine/backend): rename contract module to contracts
Closes #899
