+++ projects/products/unstable/code_forge_engine/backend/src/generate/file_generator.rs
@@ -1,4 +1,4 @@
-use crate::contract::file_spec::FileSpec;
+use crate::contracts::file_spec::FileSpec;
 use crate::diagnostics::backend_error::BackendError;
 use crate::generate::template_engine::TemplateEngine;
